% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agronomic_analysis.R
\name{harvest_index}
\alias{harvest_index}
\title{Harvest Index Calculation}
\usage{
harvest_index(grain_yield, straw_yield, verbose = TRUE)
}
\arguments{
\item{grain_yield}{Grain/economic yield}

\item{straw_yield}{Straw/stover yield}

\item{verbose}{Logical. If TRUE (default), prints formatted output to console.}
}
\value{
Data frame with harvest indices
}
\description{
Calculates Harvest Index and related partitioning indices.
}
\examples{
harvest_index(grain_yield = c(1200, 1350, 1100), straw_yield = c(2400, 2500, 2300))

}
\author{
Lalit Kumar Rolaniya, Ram Lal Jat, Monika Punia, Raja Ram Choudhary
}
